pytestProject/requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
PyMySQL==0.9.3
pytest==7.3.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
